Temp hack for ubuntu 12.10
This change makes gmail-indicator on Ubuntu 12.10 fall back to a
standard App tray indicator rather than Unity's preferred Message
indicator.

At Ubuntu 11.10, Canonical introduced a new libindicator API with Unity
which was incompatible with everything else and AFAIK, in their infinite
wisdom they have now changed it all again with a new messaging menu api
for 12.10. The previous libindicator api as used by gmail-indicator for
Unity does not seem to be supported anymore.

So using gmail-indicator is ugly on 12.10 Unity because you see 2 mail
icons in the tray but given that 12.10 has the ability to run a gmail
webapp natively that appears a superior solution than gmail-indicator
anyhow.