Update to Ubuntu 12.10 and GTK+ 3
Ubuntu have dropped their libindicate library for 12.10 (although it was
only introduced at 11.10!) and replaced it with a new libmessaging-menu
library. The new library offers no backwards compatability to
libindicate or Ubuntu 12.04. Also, gnome-shell does not seem to support
the ubuntu appindicator library as of Ubuntu 12.10 (GS3.6).

So this change upgrades gmail-indicator to libmessaging-menu for Unity
and to the GTK+ 3 introspection api for gnome-shell. Unfortunately, this
version is not supported on Ubuntu 12.04.

I have also removed the gnome-shell extension as GS 3.6 does not support it.